refactor GetFor limit to be an int to protect it from invalid fractional values

+++ src/main/java/com/zimmeren/checkout/BuyFor.java
@@ -4,21 +4,21 @@ public class BuyFor extends Special{
 
     public float buy;
     public float forTotal;
-    public float limit;
+    public int limit;
 
-    private BuyFor(float buy, float forTotal, float limit) {
+    private BuyFor(float buy, float forTotal, int limit) {
         this.buy = buy;
         this.forTotal = forTotal;
         this.limit = limit;
     }
 
-    public static BuyFor of(float buy, float forTotal, float limit) {
+    public static BuyFor of(float buy, float forTotal, int limit) {
         return new BuyFor(buy, forTotal, limit);
     }
 
     public float priceOf(float quantity, float stickerPrice) {
         float price = 0f;
-        float rounds = 0f;
+        int rounds = 0;
         while (quantity > 0f) {
             if (quantity < buy) {
                 price += quantity * stickerPrice;
